Application/users/requirements.txt
flask-login==0.6.2
gunicorn==20.1.0
pytz==2024.1
Flask==2.0.3
Werkzeug==2.0.3
boto3==1.26.70
flask-socketio==5.3.2
gunicorn==20.1.0
flask-bcrypt==1.0.1
python-dotenv==1.0.0